backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-filter==25.2
djangorestframework==3.17.1
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1